api/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
boto3==1.34.114
botocore==1.34.162
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.8
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==1.26.4
openai==1.99.3
orjson==3.11.1
pandas==2.3.1
pgvector==0.4.0
psycopg2-binary==2.9.10
pydantic==2.8.1
pydantic-settings==2.8.1
pydantic_core==2.20.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.3
rich==14.1.0
rich-toolkit==0.14.9
s3transfer==0.10.4
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
tiktoken==0.10.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
